chore(camera_test): migrate unit test
