+++ tests/github/camera_test.py
@@ -1,11 +1,36 @@
 """Unit tests for the camera module."""
 
 import pytest
+import vtk
 from ladybug_geometry.geometry3d.pointvector import Point3D
 from honeybee_vtk.camera import Camera
 from honeybee_vtk.model import Model
 from honeybee_vtk.vtkjs.schema import SensorGridOptions
 from honeybee_vtk.actor import Actor
+
+
+def test_to_vtk():
+    "Test if default properties of the camera object."
+
+    # Initialize a camera object and assess all the default properties
+    camera = Camera()
+    assert camera.identifier[0:6] == 'camera'
+    assert camera.position == (0, 0, 100)
+    assert camera.direction == (0, 0, -1)
+    assert camera.up_vector == (0, 1, 0)
+    assert camera.h_size == 60
+    assert camera.v_size == 30
+    assert camera.type == 'v'
+    assert camera.flat_view_direction[(0, 0, -1)] == [2, '+']
+    assert camera.flat_view_direction[(0, 0, 1)] == [2, '-']
+    assert camera.flat_view_direction[(0, 1, 0)] == [1, '+']
+    assert camera.flat_view_direction[(0, -1, 0)] == [1, '-']
+    assert camera.flat_view_direction[(-1, 0, 0)] == [0, '-']
+    assert camera.flat_view_direction[(1, 0, 0)] == [0, '+']
+
+    # Assess type of the outcome of the to_vtk method
+    camera = camera.to_vtk()
+    assert isinstance(camera, vtk.vtkCamera)
 
 
 def test_no_view():
